adding support to run price scheme without txn splitting

+++ hostNodePriceScheme.h
@@ -23,7 +23,7 @@ class hostNodePriceScheme : public hostNodeBase {
         // message generators
         virtual routerMsg *generateTriggerPriceUpdateMessage();
         virtual routerMsg *generatePriceUpdateMessage(double nLocal, double serviceRate, double arrivalRate, 
-                int queueSize, int reciever);
+                double queueSize, int reciever);
         virtual routerMsg *generateTriggerPriceQueryMessage();
         virtual routerMsg *generatePriceQueryMessage(vector<int> route, int routeIndex);
         virtual routerMsg *generateTriggerTransactionSendMessage(vector<int> route,
